requirements.txt
###### Requirements: pip install -r requirements.txt ######`
args==0.1.0
emoji==0.6.0
markdown2==2.3.9
prompt-toolkit==1.0.14
pyfiglet==0.8.post1
Pygments==2.7.2
PyInquirer==1.0.3
python-i18n==0.3.9
PyYAML==5.3.1
regex==2020.10.23
shortuuid==1.0.1
six==1.15.0
wcwidth==0.2.5
yattag==1.14.0